feat: map request and response
